backend/requirements.txt
fastapi
uvicorn
langchain
langchain-openai
langchain-groq
langfuse
python-dotenv
pymongo[srv]
